refactor auth pages for better reusablity and into a useAuth hook
